Renaming blog folder for datastore samples to be a proper package name, increasing coverage
