+++ blog/introduction_to_data_models_in_cloud_datastore/test_wiki.py
@@ -1,0 +1,22 @@
+# Copyright 2015, Google, Inc.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+from .wiki import main
+from tests import CloudBaseTest
+
+
+class WikiTestCase(CloudBaseTest):
+    """Simple test case that ensures the wiki code doesn't throw any errors."""
+
+    def test_main(self):
+        main(self.constants['projectId'])
