+++ blog/introduction_to_data_models_in_cloud_datastore/wiki.py
