+++ tox.ini
@@ -14,10 +14,10 @@ deps =
     coverage
     nose-exclude
 coverargs =
-        --with-coverage
-        --cover-tests
-        --cover-branches
-        --cover-inclusive
+    --with-coverage
+    --cover-tests
+    --cover-branches
+    --cover-inclusive
 
 [testenv:gae]
 deps =
@@ -49,11 +49,7 @@ deps =
     gcloud
 commands = 
     nosetests \
-        --exclude-dir=bigquery/tests/appengine \
-        --exclude-dir=bigquery/samples/appengine_auth \
         --exclude-dir=appengine \
-        --exclude-dir=datastore/ndb \
-        --exclude-dir=localtesting \
         {[testenv]coverargs} \
         {posargs}
